add test networks and refactor some java code
