+++ cpp/src/pypowsybl.h
@@ -145,6 +145,10 @@ JavaHandle createFourSubstationsNodeBreakerNetwork();
 
 bool updateSwitchPosition(const JavaHandle& network, const std::string& id, bool open);
 
+JavaHandle createBatteryNetwork();
+
+JavaHandle createDanglingLineNetwork();
+
 bool updateConnectableStatus(const JavaHandle& network, const std::string& id, bool connected);
 
 std::vector<std::string> getNetworkElementsIds(const JavaHandle& network, element_type elementType, const std::vector<double>& nominalVoltages,
